fix avg rating calculation bug

+++ app/src/main/java/com/example/read/util/calculateAvg.kt
@@ -9,7 +9,7 @@ fun calculateAvg(bookList: List<BookFB>): Double {
     }
     if (bookList.isEmpty())
         return 0.0
-    val num = (ratingSum / bookList.size).toDouble()
+    val num = (ratingSum.toDouble() / bookList.size.toDouble())
     return String.format("%.2f", num).toDouble()
 }
